Reduce number of places search is described
This removes the unclear split in code between places search was being described.
